refactor: upgrade the usePoolReward hook to be able to fetch both reward of pool or solo
