+++ packages/extension-polkagate/src/hooks/index.ts
@@ -118,7 +118,6 @@ export { default as usePoolBalances2 } from './usePoolBalances2';
 export { default as usePoolConst } from './usePoolConst';
 export { default as usePoolConsts } from './usePoolConsts';
 export { usePoolMembers } from './usePoolMembers';
-export { default as usePoolRewards } from './usePoolRewards';
 export { default as usePools } from './usePools';
 export { default as usePools2 } from './usePools2';
 export { default as usePoolStakingInfo } from './usePoolStakingInfo';
@@ -148,6 +147,7 @@ export { default as useStakingConsts2 } from './useStakingConsts2';
 export { default as useStakingRewardDestinationAddress } from './useStakingRewardDestinationAddress';
 export { default as useStakingRewards } from './useStakingRewards';
 export { default as useStakingRewards2 } from './useStakingRewards2';
+export { default as useStakingRewards3 } from './useStakingRewards3';
 export { default as useStashId } from './useStashId';
 export { default as useStashId2 } from './useStashId2';
 export { default as useTeleport } from './useTeleport';
